refactor: Internal refactoring and type fixes.
