+++ jsconfig.json
@@ -6,7 +6,7 @@
     "noEmit": true,
     "module": "node16",
     "moduleResolution": "node16",
-    "target": "es6",
+    "target": "es2022",
     "skipLibCheck": true
   }
 }
