+++ src/util/has.js
@@ -1,5 +1,1 @@
-const { hasOwnProperty } = Object.prototype;
-
-export default function(object, property) {
-  return hasOwnProperty.call(object, property);
-}
+export default Object.hasOwn;
